Add command to display full diary for a task.
Added "task <task> diary" option to display full diary for the specified
task. Also altered command so "current" can be used as an alias for the
currently active task in all instances of this command. If you have a
real task with the name "current", you're out of luck as the alias takes
precedence. Using "current" when there is no active task results in an
error.