Parsing and matching in cmdparser complete.
The cmdparser library has now had various updates for correctness and a
set of unit tests has been added to exercise it. The current stage of
development is that cmdparser can parse command specification strings
and build a parse tree, which can be used for command matching and field
extraction, as well as command-line completion.

The next step is to add a base class for "Cmd" instances which scans for
methods of of the form command_name() and uses them to build the actual
callback functions that the "cmd" module looks for. So, the docstring of
the command_name() function generates the command specification and also
becomes the help of the new command. The do_name() and complete_name()
functions are automatically added to the class within the constructor.