Added decorators for cmd.Cmd parsing.
Initial attempt at integrating cmdparser with the cmd module. Two new
decorators have been added the cmdparser, one for use with derived
cmd.Cmd classes and one for use with the methods in that class. They're
designed to wrap around the do_X() methods which implement the commands
and add parsing and completion to them.

Early indications are that matching seems to work, at least basically,
but completion isn't doing anything right now. There are not yet any
unit tests for these changse - the unit test changes committed here are
due to the change in interface added to allow code to detect the
position in the command-line where a match failure occurred.