Finished cmd integration in cmdparser.
It's now possible to use the class and method decorator in cmdparser to
implement instances derived from cmd.Cmd which use cmdparser to check
syntax of commands and extract argument values. Command-line completion
is automatically inserted, and method docstrings are reflowed to remove
the leading whitespace of the first non-initial non-blank docstring
line.

Note: Currently this leading initial indent is always removed, even if
it's not entirely whitespace - I should probably fix this, or change the
code to calculate the minimum leading whitespace across all non-blank
lines except the initial one. Currently the heuristic works for the way
I indent all my docstrings, however, so the changes can wait.