Added token factories to cmdparser's cmd support.
Updated cmdparser so that the class and method decorators are now
classes which can potentially take arguments. The only argument
currently used is to provide a token factory to the method decorator.

Also added a "context" argument to the matching functions. In the case
of the cmd integration code, this is always set to the instance of the
application-provided class derived from cmd.Cmd, so it can get real-time
context information from it for completion and matching purposes. Any
application which was using the underlying parsing code directly could
use this parameter for its own purposes - all builtin ParseItem classes
ignore it except to pass it to any children unmodified.