Case:5 - Add time aliases, info cmd and "now" time
Anywhere where a time is accepted in ttrack it should be possible to use
one of a set of tokens which represent the time at which various events
occurred. The tokens are only available if defined, but once set they're
persistent until replaced. The currently available set includes:

  - startup (time ttrack was last started)
  - shutdown (time ttrack was last quit)
  - taskstart (time any task was last started)
  - taskstop (time any task was last stopped)
  - diaryentry (time a diary entry was last entered)
  - todoadded (time a todo was last added to any task)
  - tododone (time any todo was last marked as done)

To implement these tags a new "info" table has been added to the
database schema, and I took the opportunity to also add a schema version
to this table - this will be useful as and when any incompatible changes
to the schema need to be made. This version should NOT be incremented
for compatible changes (i.e. if a new table is silently added at
startup, or a new column with a sensible default value is added).

In a related change a new "info" command can be used to view the current
version of ttrack, some summary statistics about the database and a list
of the available time aliases and their values. This command should be
updated for any other items added to the "info" database table in the
future.

Finally, a small change to the datetimeparse module to add "now" as an
available time of day.