Make package suitable for distribution
This commit moves files around into a more organised fashion suitable
for use with distutils setup scripts, and also adds a few required
files to allow unit test discovery and to support distutils.

The repository is now split into two sections as cmdparser is
potentially useful as a stand-alone library. Perhaps tracklib may be
one day, but currently it's much more closely tied to ttrack itself.

Note that all the READMEs are now in restructured text format, and the
top-level README is simply a link to the ttrack one.