Fix #5: Accept bare days of week in DateSubtree
This issue fixes a bug where the parse tree for `DateSubtree` would
accept bare days of the week (e.g. `Wednesday`) but the code to convert
the text to a `datetime` would raise an exception. The interpretation of
such a phrase is to assume an implicit `this week` after the weekday.