Fix #8: Added lastseen_time tag
To cope with cases where the "shutdown" time isn't set properly (e.g. if
a process is killed hard), this commit adds an additional lastseen_time
which can be set regularly by a client application while it's running.
On startup, if the lastseen_time is later than shutdown, the latter is
replaced with the value fo the former.