Fix #1: Added warning for long-running tasks
If the current task has been running for more than 8 consecutive hours,
ttrack will now display a warning of this fact. The check is performed
once at startup and again every time a command is entered. The warning
will only be displayed once for a given task until ttrack is restarted
or until the task is changed.