Fix #10: `delete` can now remove log entries
This commit includes changes to allow the `delete` command to remove
just log entries, and not schema details (i.e. tasks and tags). Diary
and todo entries associated with the task log entries will also be
removed. Note that todo entries are associated based on completion time,
so uncompleted todos will never be deleted unless their associated task
is deleted.

This commit also expands the range of dates accepted by the
`PastCalendarPeriodSubtree` class in datetimeparse so that it accepts
entries of the form `before <date>` and `after <date>`, where the
missing start/end date is Jan 1 1970 in the first case and the current
date in the second.