Fix #12: Added estimates, due dates, completion
This commit adds the ability for users to set an estimated duration and
due date to a task, and also allows a task to be marked as completed.
The new commands "show pending/completed" allows these stats to be
viewed. The new command "task <task> estimate/due" allows the estimate
and due date to be set, and "stop complete" can be used instead of
regular "stop" to also mark a task as completed.